# backend/utils/openapi.json
{"openapi": "3.1.0", "info": {"title": "fastapi-app", "version": "0.1.0"}, "paths": {"/api/demo/": {"get": {"tags": ["demo"], "summary": "Index", "operationId": "index", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/api/demo/db_test": {"get": {"tags": ["demo"], "summary": "Db Test", "operationId": "db_test", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/api/demo/redis_test": {"get": {"tags": ["demo"], "summary": "Redis Test", "operationId": "redis_test", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/api/demo/{demo_id}": {"get": {"tags": ["demo"], "summary": "Show", "operationId": "show", "parameters": [{"required": true, "schema": {"type": "string", "title": "Demo Id"}, "name": "demo_id", "in": "path"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/auth/token": {"post": {"tags": ["auth"], "summary": "Token", "description": "\u7528\u6237\u540d+\u5bc6\u7801\u767b\u5f55", "operationId": "token", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/OAuth2PasswordRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Token"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/auth/cellphone/token": {"post": {"tags": ["auth"], "summary": "Cellphone Token", "description": "\u624b\u673a\u53f7+\u9a8c\u8bc1\u7801\u767b\u5f55", "operationId": "cellphone_token", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/OAuth2CellphoneRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Token"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/auth/cellphone/verification_code": {"post": {"tags": ["auth"], "summary": "Send Verification Code", "description": "\u53d1\u9001\u9a8c\u8bc1\u7801", "operationId": "send_verification_code", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Body_auth-send_verification_code"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/users/me": {"get": {"tags": ["users"], "summary": "Me", "description": "\u5f53\u524d\u767b\u5f55\u7528\u6237\u4fe1\u606f", "operationId": "me", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserDetail"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/api/users/add_user": {"post": {"tags": ["users"], "summary": "Add User", "operationId": "add_user", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Body_users-add_user"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/api/chats/new": {"get": {"tags": ["chats"], "summary": "New Session", "operationId": "new_session", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ChatBase"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/api/chats/del": {"get": {"tags": ["chats"], "summary": "Del Session", "operationId": "del_session", "parameters": [{"required": true, "schema": {"type": "string", "title": "Session Id"}, "name": "session_id", "in": "query"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/api/chats/get": {"get": {"tags": ["chats"], "summary": "Get Session", "operationId": "get_session", "parameters": [{"required": true, "schema": {"type": "string", "title": "Session Id"}, "name": "session_id", "in": "query"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ChatBase"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/api/chats/list": {"get": {"tags": ["chats"], "summary": "List Session", "operationId": "list_session", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/ChatBase"}, "type": "array", "title": "Response Chats-List Session"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/api/chats/ask": {"post": {"tags": ["chats"], "summary": "Ask", "operationId": "ask", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ChatRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/api/chats/set_sys_prompt": {"post": {"tags": ["chats"], "summary": "Set Sys Prompt", "operationId": "set_sys_prompt", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Body_chats-set_sys_prompt"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/api/chats/get_sys_prompt": {"get": {"tags": ["chats"], "summary": "Get Sys Prompt", "operationId": "get_sys_prompt", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/api/chats/set_cookies": {"post": {"tags": ["chats"], "summary": "Set Cookies", "operationId": "set_cookies", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Body_chats-set_cookies"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/api/chats/events": {"get": {"tags": ["chats"], "summary": "Get Events", "operationId": "get_events", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/items/": {"post": {"tags": ["items"], "summary": "Create Item", "operationId": "create_item", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Item"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ResponseMessage"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/items/": {"get": {"tags": ["items"], "summary": "Get Items", "operationId": "get_items", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/Item"}, "type": "array", "title": "Response Items-Get Items"}}}}}}}, "/users/": {"post": {"tags": ["users"], "summary": "Create User", "operationId": "create_user", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/User"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ResponseMessage"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"Body_auth-send_verification_code": {"properties": {"cellphone": {"type": "string", "title": "Cellphone"}}, "type": "object", "required": ["cellphone"], "title": "Body_auth-send_verification_code"}, "Body_chats-set_cookies": {"properties": {"cookies": {"type": "string", "title": "Cookies"}}, "type": "object", "required": ["cookies"], "title": "Body_chats-set_cookies"}, "Body_chats-set_sys_prompt": {"properties": {"name": {"type": "string", "title": "Name"}, "content": {"type": "string", "title": "Content"}}, "type": "object", "required": ["name", "content"], "title": "Body_chats-set_sys_prompt"}, "Body_users-add_user": {"properties": {"username": {"type": "string", "title": "Username"}, "password": {"type": "string", "title": "Password"}}, "type": "object", "required": ["username", "password"], "title": "Body_users-add_user"}, "ChatBase": {"properties": {"id": {"type": "integer", "title": "Id"}, "title": {"type": "string", "title": "Title"}, "user_id": {"type": "string", "title": "User Id"}, "session_id": {"type": "string", "title": "Session Id"}, "message": {"type": "string", "title": "Message"}}, "type": "object", "required": ["id", "title", "user_id", "session_id"], "title": "ChatBase"}, "ChatRequest": {"properties": {"content": {"type": "string", "title": "Content"}, "session_id": {"type": "string", "title": "Session Id"}, "stream": {"type": "boolean", "title": "Stream", "default": false}}, "type": "object", "required": ["content"], "title": "ChatRequest"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "Item": {"properties": {"name": {"type": "string", "title": "Name"}, "price": {"type": "number", "title": "Price"}}, "type": "object", "required": ["name", "price"], "title": "Item"}, "OAuth2CellphoneRequest": {"properties": {"grant_type": {"type": "string", "pattern": "cellphone", "title": "Grant Type"}, "cellphone": {"type": "string", "title": "Cellphone"}, "verification_code": {"type": "string", "title": "Verification Code"}, "scope": {"type": "string", "title": "Scope", "default": ""}, "client_id": {"type": "string", "title": "Client Id"}, "client_secret": {"type": "string", "title": "Client Secret"}}, "type": "object", "required": ["cellphone", "verification_code"], "title": "OAuth2CellphoneRequest"}, "OAuth2PasswordRequest": {"properties": {"grant_type": {"type": "string", "pattern": "password", "title": "Grant Type"}, "username": {"type": "string", "title": "Username"}, "password": {"type": "string", "title": "Password"}, "scope": {"type": "string", "title": "Scope", "default": ""}, "client_id": {"type": "string", "title": "Client Id"}, "client_secret": {"type": "string", "title": "Client Secret"}}, "type": "object", "required": ["username", "password"], "title": "OAuth2PasswordRequest", "description": "\u53c2\u7167 fastapi.security.OAuth2PasswordRequestForm\uff0c\u628a\u8bf7\u6c42\u4f53\u4eceform\u6539\u4e3ajson\u683c\u5f0f"}, "ResponseMessage": {"properties": {"message": {"type": "string", "title": "Message"}}, "type": "object", "required": ["message"], "title": "ResponseMessage"}, "Token": {"properties": {"token_type": {"type": "string", "title": "Token Type"}, "expires_in": {"type": "integer", "title": "Expires In"}, "access_token": {"type": "string", "title": "Access Token"}}, "type": "object", "required": ["token_type", "expires_in", "access_token"], "title": "Token"}, "User": {"properties": {"username": {"type": "string", "title": "Username"}, "email": {"type": "string", "title": "Email"}}, "type": "object", "required": ["username", "email"], "title": "User"}, "UserDetail": {"properties": {"id": {"type": "integer", "title": "Id"}, "username": {"type": "string", "title": "Username"}, "nickname": {"type": "string", "title": "Nickname"}, "gender": {"type": "string", "title": "Gender"}, "avatar": {"type": "string", "title": "Avatar"}, "cellphone": {"type": "string", "title": "Cellphone"}, "email": {"type": "string", "title": "Email"}, "email_verified_at": {"type": "string", "format": "date-time", "title": "Email Verified At"}, "state": {"type": "string", "title": "State"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}}, "type": "object", "required": ["id", "username", "nickname", "gender", "avatar", "state", "created_at"], "title": "UserDetail"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}, "securitySchemes": {"OAuth2PasswordBearer": {"type": "oauth2", "flows": {"password": {"scopes": {}, "tokenUrl": "token"}}}}}}